requirements.txt
altair==5.5.0
altgraph==0.17.4
asttokens==3.0.1
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.3
cryptography==41.0.7
debugpy==1.8.19
decorator==4.4.2
et_xmlfile==2.0.0
executing==2.2.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keyboard==0.13.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mido==1.3.3
moviepy==1.0.3
narwhals==1.19.1
nest-asyncio==1.6.0
numpy==2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.0
parso==0.8.5
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.5.1
pretty_midi==0.2.10
proglog==0.1.12
prompt_toolkit==3.0.52
protobuf==5.29.2
psutil==7.2.1
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.21
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
python-dateutil==2.8.2
python-dotenv==1.2.1
python-vlc==3.0.21203
pytz==2024.1
pywin32-ctypes==0.2.3
pyzmq==27.1.0
referencing==0.35.1
reportlab==4.4.3
requests==2.31.0
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.4.0
scipy==1.12.0
setuptools==75.6.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.41.1
tenacity==9.0.0
threadpoolctl==3.2.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.1.0
wakeonlan==3.1.0
watchdog==6.0.0
wcwidth==0.3.1
websockets==12.0
